Backend/requirement.txt
fastapi
uvicorn
torch
torchvision
mtcnn
opencv-python
pillow
